created header file for user_output_query.cpp file
