+++ data/output/abp_query_output.csv
@@ -1,27 +1,35 @@
 time,value,port,component
 00:00:20:000,1,packet_sent_out,sender1
 00:00:20:000,11,data_out,sender1
+00:00:33:000,1,out,receiver1
 00:00:36:000,1,ack_received_out,sender1
 00:00:46:000,2,packet_sent_out,sender1
 00:00:46:000,20,data_out,sender1
+00:00:59:000,0,out,receiver1
 00:01:02:000,0,ack_received_out,sender1
 00:01:12:000,3,packet_sent_out,sender1
 00:01:12:000,31,data_out,sender1
+00:01:25:000,1,out,receiver1
 00:01:28:000,1,ack_received_out,sender1
 00:01:40:000,4,packet_sent_out,sender1
 00:01:40:000,40,data_out,sender1
+00:01:53:000,0,out,receiver1
 00:01:56:000,0,ack_received_out,sender1
 00:02:06:000,5,packet_sent_out,sender1
 00:02:06:000,51,data_out,sender1
+00:02:19:000,1,out,receiver1
 00:02:22:000,1,ack_received_out,sender1
 00:20:10:000,1,packet_sent_out,sender1
 00:20:10:000,11,data_out,sender1
+00:20:23:000,1,out,receiver1
 00:20:26:000,1,ack_received_out,sender1
 00:20:36:000,2,packet_sent_out,sender1
 00:20:36:000,20,data_out,sender1
+00:20:49:000,0,out,receiver1
 00:20:52:000,0,ack_received_out,sender1
 00:21:02:000,3,packet_sent_out,sender1
 00:21:02:000,31,data_out,sender1
 00:21:32:000,3,packet_sent_out,sender1
 00:21:32:000,31,data_out,sender1
+00:21:45:000,1,out,receiver1
 00:21:48:000,1,ack_received_out,sender1
